requirements.txt
torch
torchvision
numpy
matplotlib
Pillow
streamlit-image-select>=0.6.0
st-clickable-images